// src/app/models/forcast-input.model.ts
export interface ForecastInput {
  intl_Predict60_noOfFlight_InBound: number;
  intl_Predict60_noOfFlight_OutBound: number;
  intl_Predict120_noOfFlight_InBound: number;
  intl_Predict120_noOfFlight_OutBound: number;
  intl_Predict60_noOfPax_InBound: number;
  intl_Predict60_noOfPax_OutBound: number;
  intl_Predict120_noOfPax_InBound: number;
  intl_Predict120_noOfPax_OutBound: number;
  crossStrait_Predict60_noOfFlight_InBound: number;
  crossStrait_Predict60_noOfFlight_OutBound: number;
  crossStrait_Predict120_noOfFlight_InBound: number;
  crossStrait_Predict120_noOfFlight_OutBound: number;
  crossStrait_Predict60_noOfPax_InBound: number;
  crossStrait_Predict60_noOfPax_OutBound: number;
  crossStrait_Predict120_noOfPax_InBound: number;
  crossStrait_Predict120_noOfPax_OutBound: number;
  domestic_Predict60_noOfFlight_InBound: number;
  domestic_Predict60_noOfFlight_OutBound: number;
  domestic_Predict30_noOfFlight_InBound: number;
  domestic_Predict30_noOfFlight_OutBound: number;
  domestic_Predict60_noOfPax_InBound: number;
  domestic_Predict60_noOfPax_OutBound: number;
  domestic_Predict30_noOfPax_InBound: number;
  domestic_Predict30_noOfPax_OutBound: number;
}